create product service, create function for save data product in service, implements unit test for function saveProduct in service product

+++ domain/product.go
@@ -3,9 +3,9 @@ package domain
 type Product struct {
 	ProductId          int32  `gorm:"primaryKey"`
 	ProductName        string `db:"product_name"`
-	CategoryId         int32  `db:"category_id"`
+	CategoryId         int64  `db:"category_id"`
 	Price              int64  `db:"price"`
-	Stock              int32  `db:"stock"`
+	Stock              int64  `db:"stock"`
 	ProductDescription string `db:"product_description"`
 }
